requirements.txt
flet
google-generativeai
supabase
fpdf